(GA) default single precision

+++ mastsel/mavisPsf.py
@@ -12,8 +12,8 @@ from mastsel.mavisFormulas import *
 
 fit_window_max_size = 512
 defaultArrayBackend = cp
-defaultArrayDtype = defaultArrayBackend.float64
-defaultArrayCDtype = defaultArrayBackend.complex128
+defaultArrayDtype = defaultArrayBackend.float32
+defaultArrayCDtype = defaultArrayBackend.complex64
 
 def mastselPsfPrecision(precision=None, dtype=None):
     global defaultArrayDtype, defaultArrayCDtype
